+++ src/main/java/wooteco/subway/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.response;
 
 public class ErrorResponse {
     private String message;
